refactor OgpPreviewCard
